changed to BigDecimal for storing money

+++ NetPay/src/main/java/com/sonalake/task/NetPayResource.java
@@ -1,5 +1,7 @@
 package com.sonalake.task;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -10,7 +12,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class NetPayResource {
-	private Float netPay;
+	private BigDecimal netPay;
 	private String countryCode;
 	private String currencyCode;
